Get Response Modified

+++ inner_department_communication-backend/src/main/java/com/example/inner_department_communication_backend/repo/ProjectRepository.java
@@ -50,6 +50,6 @@ Optional<Project> findById(Long id);
 
 @Transactional
 @Query(value = "SELECT p.* FROM project p where project_manager_id=?1", nativeQuery = true)
-public List<Project> findProjectManagerId(int id);
+public List<Project> findProjectManagerId(Long id);
 
 }
